backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
google-generativeai==0.11.0
textblob==0.17.1
gunicorn==20.1.0